Fixed QM scripts for Abyssea-Misareaux
