+++ scripts/zones/Abyssea-Misareaux/npcs/qm15.lua
@@ -1,32 +1,27 @@
 -----------------------------------
 -- Zone: Abyssea-Misareaux
 --  NPC: qm15 (???)
--- Spawns Sirrush
+-- Spawns Sobek
 -- @pos ? ? ? 216
 -----------------------------------
+require("scripts/globals/keyitems");
 require("scripts/globals/status");
-
------------------------------------
--- onTrade Action
------------------------------------
-
-function onTrade(player,npc,trade)
---[[
-    if (trade:hasItemQty(3086,1) and trade:getItemCount() == 1) then -- Player has all the required items.
-        if (GetMobAction(17662465) == ACTION_NONE) then -- Mob not already spawned from this
-            SpawnMob(17662465):updateClaim(player); -- Spawn NM, Despawn after inactive for 5 minutes (pt has to reclaim within 5 of a wipe)
-            player:tradeComplete();
-        end
-    end
-]]
-end;
 
 -----------------------------------
 -- onTrigger Action
 -----------------------------------
 
 function onTrigger(player,npc)
-    player:startEvent(1010, 3086); -- Inform player what items they need.
+--[[
+    if (GetMobAction(17662478) == ACTION_NONE) then -- NM not already spawned from this
+        if (player:hasKeyItem(BLOODSTAINED_BUGARD_FANG) and player:hasKeyItem(GNARLED_LIZARD_NAIL)
+        and player:hasKeyItem(MOLTED_PEISTE_SKIN)) then -- I broke it into 3 lines at the 'and' because it was so long.
+            player:startEvent(1021, BLOODSTAINED_BUGARD_FANG, GNARLED_LIZARD_NAIL, MOLTED_PEISTE_SKIN); -- Ask if player wants to use KIs
+        else
+            player:startEvent(1020, BLOODSTAINED_BUGARD_FANG, GNARLED_LIZARD_NAIL, MOLTED_PEISTE_SKIN); -- Do not ask, because player is missing at least 1.
+        end
+    end
+]]
 end;
 
 -----------------------------------
@@ -34,8 +29,8 @@ end;
 -----------------------------------
 
 function onEventUpdate(player,csid,option)
-    -- printf("CSID: %u",csid);
-    -- printf("RESULT: %u",option);
+    -- printf("CSID2: %u",csid);
+    -- printf("RESULT2: %u",option);
 end;
 
 -----------------------------------
@@ -45,4 +40,10 @@ end;
 function onEventFinish(player,csid,option)
     -- printf("CSID: %u",csid);
     -- printf("RESULT: %u",option);
+    if (csid == 1021 and option == 1) then
+        SpawnMob(17662478):updateClaim(player); -- Spawn NM, Despawn after inactive for 5 minutes (pt has to reclaim within 5 of a wipe)
+        player:delKeyItem(BLOODSTAINED_BUGARD_FANG);
+        player:delKeyItem(GNARLED_LIZARD_NAIL);
+        player:delKeyItem(MOLTED_PEISTE_SKIN);
+    end
 end;
