+++ scripts/zones/Abyssea-Misareaux/npcs/qm9.lua
@@ -12,7 +12,7 @@ require("scripts/globals/status");
 
 function onTrade(player,npc,trade)
 --[[
-    if (trade:hasItemQty(3093,1) and trade:getItemCount() == 1) then -- Player has all the required items.
+    if (trade:hasItemQty(3093,1) and trade:hasItemQty(3094,1) and trade:getItemCount() == 2) then -- Player has all the required items.
         if (GetMobAction(17662472) == ACTION_NONE) then -- Mob not already spawned from this
             SpawnMob(17662472):updateClaim(player); -- Spawn NM, Despawn after inactive for 5 minutes (pt has to reclaim within 5 of a wipe)
             player:tradeComplete();
@@ -26,7 +26,7 @@ end;
 -----------------------------------
 
 function onTrigger(player,npc)
-    player:startEvent(1010, 3093); -- Inform player what items they need.
+    player:startEvent(1010, 3093, 3094); -- Inform player what items they need.
 end;
 
 -----------------------------------
